Conflicts:
	app/views/rooms/index.html.haml
	app/views/rooms/show.html.haml

+++ app/models/session.rb
@@ -6,13 +6,17 @@ class Session < ActiveRecord::Base
       time = time.split.inject { |count, unit| count.to_i.send(unit) }
     end
  
-    delete_all "updated_at < '#{time.ago.to_s(:db)}'"
+    destroy_all "updated_at < '#{time.ago.to_s(:db)}'"
   end
   
   def check_if_stay_paid
-    s= Stay.find self[:stay_id]
-    if !s.paid?
-      s.destroy
+    if self[:stay_id].nil?
+      return true
+    else
+      s= Stay.find self[:stay_id]
+      if !s.paid? && agency_id.nil?
+        s.destroy
+      end
     end
   end
 end
